Expermineting with support for a secondary menu
This changeset adds an experimental "alt menu", allowing us to begin
adding some per-item actions without cluttering the main user interface
with stuff that is going to be irrelevant 90% of the time. The
appearance and functionality of these alt menu behaviors is subject to
change! The secondary menu

As a useful proof of concept, a "delete" action has been provided to
every clipping.